refactor: move popup menu to components
