+++ src/components/ProjectDashboard/components/UpdatesPanel/components/ProjectUpdate.tsx
@@ -9,7 +9,7 @@ import { useIsUserAddress } from 'hooks/useIsUserAddress'
 import { useWalletSignIn } from 'hooks/useWalletSignIn'
 import { useCallback, useContext, useMemo } from 'react'
 import { emitErrorNotification } from 'utils/notifications'
-import { PopupMenu } from '../../PopupMenu/PopupMenu'
+import { PopupMenu } from '../../../../ui/PopupMenu'
 import { ProjectUpdate as ProjectUpdateEntity } from '../../ProjectUpdatesProvider'
 import { ProjectUpdatesContext } from '../../ProjectUpdatesProvider/ProjectUpdatesProvider'
 import { useFactoredProjectId } from '../hooks/useFactoredProjectId'
